refactor: customer 패키지 이동 및 불필요한 부분 제거 & 메서드 이름, 위치 수정
